requirements.txt
certifi==2024.6.2
charset-normalizer==3.3.2
idna==3.7
pyotp==2.9.0
python-dotenv==1.0.1
requests==2.32.3
setuptools==70.1.1
synology-api==0.7.2
urllib3==2.2.2
